competitor-analysis-sdk-migration and change all references to Mino into Tinyfish agent
