+++ competitor-analysis/lib/utils.ts
@@ -5,8 +5,8 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
 
-// Mino SSE Event Types
-export interface MinoEvent {
+// Tinyfish Agent streaming event types (normalized)
+export interface TinyFishAgentEvent {
   type: "STEP" | "COMPLETE" | "ERROR" | string;
   status?: string;
   message?: string;
@@ -17,15 +17,15 @@ export interface MinoEvent {
 }
 
 /**
- * Parse an SSE line into a MinoEvent
+ * Parse an SSE line into a TinyFishAgentEvent
  */
-export function parseSSELine(line: string): MinoEvent | null {
+export function parseSSELine(line: string): TinyFishAgentEvent | null {
   if (!line.startsWith("data: ")) {
     return null;
   }
 
   try {
-    return JSON.parse(line.slice(6)) as MinoEvent;
+    return JSON.parse(line.slice(6)) as TinyFishAgentEvent;
   } catch {
     return null;
   }
@@ -34,21 +34,21 @@ export function parseSSELine(line: string): MinoEvent | null {
 /**
  * Check if event indicates successful completion
  */
-export function isCompleteEvent(event: MinoEvent): boolean {
+export function isCompleteEvent(event: TinyFishAgentEvent): boolean {
   return event.type === "COMPLETE" && event.status === "COMPLETED";
 }
 
 /**
  * Check if event indicates an error
  */
-export function isErrorEvent(event: MinoEvent): boolean {
+export function isErrorEvent(event: TinyFishAgentEvent): boolean {
   return event.type === "ERROR" || event.status === "FAILED";
 }
 
 /**
  * Format a step event into a readable message
  */
-export function formatStepMessage(event: MinoEvent): string {
+export function formatStepMessage(event: TinyFishAgentEvent): string {
   const stepInfo = event.step && event.totalSteps
     ? `[${event.step}/${event.totalSteps}]`
     : "[STEP]";
