+++ competitor-analysis/package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-separator": "^1.1.8",
     "@radix-ui/react-slot": "^1.2.4",
     "@radix-ui/react-tabs": "^1.1.13",
+    "@tiny-fish/sdk": "^0.0.7",
     "@radix-ui/react-tooltip": "^1.2.8",
     "ai": "^4.0.0",
     "class-variance-authority": "^0.7.1",
